requirements.txt
asn1crypto==1.5.1
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.17.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==309
pyzmq==26.2.1
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
stack-data==0.6.3
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13